move ValidateToken from limes.Driver to limes.AuthParameters
